requirements.txt
a2wsgi
fastapi
pyinstaller
pywebview
